refactor: move all search files over to discover
The search files that were in app-series have now been moved into app-discover
