+++ app/src/main/java/com/chesire/malime/injection/androidmodules/FragmentModule.kt
@@ -1,11 +1,11 @@
 package com.chesire.malime.injection.androidmodules
 
 import com.chesire.malime.app.discover.DiscoverFragment
+import com.chesire.malime.app.discover.search.SearchFragment
 import com.chesire.malime.app.profile.ProfileFragment
 import com.chesire.malime.app.series.detail.SeriesDetailSheetFragment
 import com.chesire.malime.app.series.list.anime.AnimeFragment
 import com.chesire.malime.app.series.list.manga.MangaFragment
-import com.chesire.malime.app.series.search.SearchFragment
 import com.chesire.malime.app.settings.oss.OssFragment
 import com.chesire.malime.app.timeline.TimelineFragment
 import com.chesire.malime.flow.login.details.DetailsFragment
